Remove deprecated stemcell lines
The Foundation has adopted a policy of removing support for stemcells
older than 3 years:
https://github.com/cloudfoundry/community/blob/1cb77d27fbcdc396d1fa290ceaf5b6aab56fc498/toc/rfc/rfc-0010-stemcell-cleanup.md?plain=1#L26

In order to follow this policy, we need to remove older stemcells from
bosh.io in order to prevent broken links.

The last Xenial stemcell was published on April 29th, 2021, and thus
passed the 3 year mark earlier this year.

Signed-off-by: Brian Upton <brian.upton@broadcom.com>